Prevent a meta leak of video orientation
The previous logic lead to meta leaks: If a video orientation header is
not present, it was quite easy to infer that the 0° orientation was in
use. This patch just always sends the video orientation, so there are
no leaks in regards to orientation when combined with RTP header
extension encryption.