New build tools
The two Docker images have been reduced to one which is significantly smaller.

The build tools image contains necessary tools to fetch the source files and
build libwebrtc from them. Source files are now stored in the webrtc folder
instead of the Docker image. This allows to easily modify the files locally.

An accompanying script (cli.sh) can be used to

- build the tools image which will then be used to
- fetch the sources
- update the sources
- patch the sources
- build specific or all Android targets
- run custom commands in the tools container

In addition, the build has been reduced to the bare minimum required for the
target and subsequent builds are sped up drastically via the use of ccache.