DTMF patch: Remove initialization in RTP Sender
Currently the `RtpSender` class in Java and the `RTCRtpSender` in ObjC
always initialize a `DtmfSender` instance, which in turn calls
`VideoRtpSender::GetDtmfSender` in `pc/rtp_sender.cc`.

That call results in an error log:

     RTC_LOG(LS_ERROR) << "Tried to get DTMF sender from video sender.";

Remove the `DtmfSender` instance from `RtpSender.java` and
`RTCRtpSender.mm`, so that the error logs disappear.